requirements.txt
Django==2.2.6
pytz==2019.3
sqlparse==0.3.0
django-redactoreditor==1.2.10
